src/app/features/assessment/components/post-list/post-list.component.ts
import { Component, OnDestroy, OnInit } from '@angular/core';
import { Store } from '@ngrx/store';
import { Subject, takeUntil } from 'rxjs';

import { PostModel } from '@/api/models/post.model';

import * as Selectors from '../../store/selectors/post.selector';

@Component({
  selector: 'app-post-list',
  templateUrl: './post-list.component.html',
  styleUrls: ['./post-list.component.scss'],
})
export class PostListComponent implements OnInit, OnDestroy {
  posts: PostModel[] = [];

  private destroy$ = new Subject<void>();

  constructor(private store: Store) {}

  ngOnInit(): void {
    this.store
      .select(Selectors.selectPosts)
      .pipe(takeUntil(this.destroy$))
      .subscribe({
        next: (posts) => (this.posts = posts),
      });
  }

  ngOnDestroy(): void {
    this.destroy$.next();
    this.destroy$.complete();
  }
}
